requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
Levenshtein==0.23.0
MarkupSafe==2.1.3
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.3
packaging==23.2
preshed==3.0.9
pydantic==2.5.3
pydantic_core==2.14.6
pyenchant==3.2.2
python-Levenshtein==0.23.0
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
strsimpy==0.2.1
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
